requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
idna==3.4
jieba==0.42.1
joblib==1.2.0
mecab-python3==1.0.6
nltk==3.8.1
numpy==1.24.2
Pillow==9.4.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
sklearn==0.0.post1
threadpoolctl==3.1.0
torch==1.13.1
torchaudio==0.13.1
torchtext==0.14.1
torchvision==0.14.1
tqdm==4.65.0
typing_extensions==4.5.0
unidic-lite==1.0.8
urllib3==1.26.14
